Front-facing documentation x 3.
This commit significantly improves the front-facing "README.md" by
emphasizing that this repository *only* officially hosts the
Lutris-based Linux installer and that Windows users should ideally
redirect Windows-specific concerns to the Beast's Lair forum. This
resolves #5 and should help to avoid similar issues in the future.

Unrelatedly, this commit also adds thoroughly untested (and thus
currently disabled) preliminary support for the next stable release of
the Ultimate Edition -- which has yet to be released, of course.