Front-facing documentation x 4.
This commit addresses #6 by revising the FAQ in the front-facing
"README.md" to exhaustively list all of the symptoms of patch
corruption, the leading cause of most startup failures. This includes
a screenshot generously captured and denoted by @CavaleriOmar at #6,
which should help to ameliorate similar issues in the future. Thanks!