Front-facing documentation x 5.
This commit addresses the last subissue that cropped up in #6 by
revising the FAQ in the front-facing "README.md" yet again to
exhaustively list all of the symptoms of bad Vulkan and DXVK
installations and misconfiguration, the next leading cause of most
startup failures. This includes log output yet again generously captured
and denoted by @CavaleriOmar at #6, which should help to ameliorate
similar issues in the future. Thanks!