Update fate-stay-night-realta-nua-ultimate-edition.yml
Spun op a VM and did a clean install with the install script adjusted to use wine 6.21-5.
It installed and launched without problems and menus worked too. I tested it until the first choice and it didn't crash and burn afterwards.
There is also a newer version of wine 6.14 available (6.14-4) but this would still result in holding back wine.

When 6.11 is a year old or when a next major version of wine comes out we can unpin it or maybe earlier. It would prevent having to keep track of wine versions as we should just expect it to work regardless of version.